Move tsx files out of /includes, then include the includes in the final build. For now.
