+++ gemp-module/gemp-stccg-client/src/main/web/src/Login/login.jsx
@@ -2,7 +2,7 @@ import { createRoot } from 'react-dom/client';
 import "../../js/jquery/jquery-3.7.1.js";
 import "../../js/jquery/jquery-ui-1.14.1/jquery-ui.js";
 import GempClientCommunication from "../../js/gemp-022/communication.js";
-import LoginRegisterTabs from "../../includes/LoginRegisterTabs.jsx";
+import LoginRegisterTabs from "./LoginRegisterTabs.jsx";
 
 
 var comm = new GempClientCommunication("/gemp-stccg-server", function () {
